stopping for now
figured out how to pass args forward to a function using a variadic template.
may need to restrict it, though. Promising a maze is in the first slot would provide
a lot of flexibility at the cost of making people pass the maze in as the first parameter
to all of their implemented algorithms.